Backend/google_map/requirements.txt
googlemaps==4.10.0
pandas>=1.3.0
python-dotenv>=0.19.0
